+++ includes/class-wordpress-starter-widget.php
@@ -1,19 +1,21 @@
 <?php
-/*
-	Copyright 2014 Michael Cannon (email: mc@aihr.us)
+/**
+Aihrus WordPress Starter
+Copyright (C) 2014  Michael Cannon
 
-	This program is free software; you can redistribute it and/or modify
-	it under the terms of the GNU General Public License, version 2, as
-	published by the Free Software Foundation.
+This program is free software; you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation; either version 2 of the License, or
+(at your option) any later version.
 
-	This program is distributed in the hope that it will be useful,
-	but WITHOUT ANY WARRANTY; without even the implied warranty of
-	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-	GNU General Public License for more details.
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
 
-	You should have received a copy of the GNU General Public License
-	along with this program; if not, write to the Free Software
-	Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+You should have received a copy of the GNU General Public License along
+with this program; if not, write to the Free Software Foundation, Inc.,
+51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
 require_once AIHR_DIR_INC . 'class-aihrus-widget.php';
@@ -27,9 +29,9 @@ class WordPress_Starter_Widget extends Aihrus_Widget {
 
 	public function __construct( $classname = null, $description = null, $id_base = null, $title = null ) {
 		$classname   = 'WordPress_Starter_Widget';
-		$description = esc_html__( 'Display WordPress Starter entries.' );
+		$description = esc_html__( 'Display Aihrus WordPress Starter entries.' );
 		$id_base     = self::ID;
-		$title       = esc_html__( 'WordPress Starter' );
+		$title       = esc_html__( 'Aihrus WordPress Starter' );
 
 		parent::__construct( $classname, $description, $id_base, $title );
 	}